requirements_ml.txt
# Core Data Science Libraries
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Machine Learning
scikit-learn>=1.4.0
umap-learn>=0.5.5

# Deep Learning & NLP
torch>=2.1.0
transformers>=4.36.0

# Text Processing
nltk>=3.8.1
spacy>=3.7.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
wordcloud>=1.9.3

# Progress Bars & Utilities
tqdm>=4.66.0
pyyaml>=6.0.1

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.1

# Existing requirements
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
